changed most data types from int to double for easier and more accurate calculations

+++ src/main/java/breakout/gameObject/Brick.java
@@ -6,7 +6,7 @@ public class Brick extends GameObject {
     /** if a block is still intact or got destroyed */
     private boolean isDestroyed;
 
-    public Brick (int x, int y, int width, int height, Color color) {
+    public Brick (double x, double y, double width, double height, Color color) {
         super(x, y, width, height, color);
         isDestroyed = false;
     }
